# optimize null, true, and false prefix checks